requirements.txt
astroid==2.4.1
asyncpg==0.21.0
attrs==20.1.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
databases==0.3.2
decorator==4.4.2
dodgy==0.2.1
fastapi==0.61.0
flake8==3.8.3
flake8-polyfill==1.0.2
gunicorn==20.0.4
h11==0.9.0
html5lib==1.1
httptools==0.1.1
idna==2.10
iniconfig==1.0.1
ipdb==0.13.3
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
multitasking==0.0.9
numpy==1.19.1
packaging==20.4
pandas==1.1.1
parso==0.7.1
pep8-naming==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.6
prospector==1.3.0
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pydantic==1.6.1
pydocstyle==5.1.0
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.2
pylint-celery==0.3
pylint-django==2.0.15
pylint-flask==0.6
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requirements-detector==0.7
setoptconf==0.2.0
six==1.15.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.19
starlette==0.13.6
toml==0.10.1
traitlets==4.3.3
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
wrapt==1.12.1
yfinance==0.1.54
